wc: return libsoup input stream
By default, EamWc downloads a file and stores it in the file-system, reporting
an error if something failed. But there's another scenario: download the file
into memory and parse it.

For this scenario, this commit enables a "hack mode", called return-instream,
where, instead of storing the file into a directory, the process ends
returning the input stream created by libsoup.