Change the ListAvailable return value
Instead of coalescing all the installable and updatable applications we
should return two disting tuples. This way the app store can store the
data separately and match the correct information.

We need to change the ListAvailable method as well as the
AvailableApplicationsChanged signal, since they both have the same
signature.