Fixed missing deps script for virtual packages
Include a hard-coded work-around to ensure we ignore a few packages
that are available in apt but not installable due to dependency issues.
In these cases, we need to ensure that we choose the alternate option
to satisfy the dependency list.

We also now allow specification of either the debian package name
or a .deb file name.

The default core is now none rather than eos-core, so that the default
is simply to list all the dependencies of a package.

Output is now sorted alphabetically.

[endlessm/eos-shell#2633]