Handle external components
The .info file in a bundle can specify one or more external files to be
downloaded, in a [External] section.

It requires three values:

- url = `url to download`
- filename = `output filename for the downloaded content`
- sha256sum = `SHA256 checksum for the downloaded content`

It is possible to define more than one file to download. In this case,
each of the parameters will be indexed, starting in 0: url0],
filename[0], sha256sum[0], url[1], filename[1], sha256sum[1], ...

All the downloaded content will be stored at
${EAM_TMP}/${APP_ID}/external

[endlessm/eos-shell#2799]