Handle external components
The .info file in a bundle can specify one or more external files to be
downloaded, in a [External] section.

It requires three values:

- url = `url to download`
- filename = `output filename for the downloaded content`
- sha256sum = `SHA256 checksum for the downloaded content`

It is possible to define more than one file to download. In this case,
we need to define several external sections, giving to each them an
identifier:

```
[External <id1>]
url=...
filename=...
sha256sum=...

[External <id2>]
url=...
filename=...
sha256sum=...
```

All the downloaded content will be stored at
${EAM_TMP}/${APP_ID}/external

[endlessm/eos-shell#2799]