scripts: Create the /endless directory structure when installing
The image builder installs apps without the app-manager daemon running,
so it needs to have the directories in EAM_PREFIX (/endless) created
another way. Add a new utils function, create_os_directories(), that
creates all the needed directories. Add new scripts in the phases where
app installation occurs to call the function.

This fixes the issue for the image builder, but it's also more robust as
it ensures that the directories exist beyond the one time when the
daemon is initialized. That code is left in place as it also handles the
attempt to create the /endless to /var/endless symlink.

[endlessm/eos-shell#2966]