scripts: Create link directories before attempting symlink
Before attempting to make a symlink in the link directory, ensure that
the link directory exists. This is most likely not needed in most cases
since the links we're creating are going in the well defined shared
directories, but this makes the code more robust in case a different
link directory is passed to symbolic_links().

[endlessm/eos-shell#2966]