scripts: Byte compile python modules on install or update
The python bytecode is not distributed with the bundles, so we need to
generate it on install or update. We can't let python recursively
compile the entire app directory since compiling a module with the wrong
python version will cause an error. Therefore, we need to narrow
compiling to the python site directories where we know the python
version.

Since updating or removing the app will delete the entire old app
directory, we don't need to worry about cleaning up any existing python
bytecode. The new app installation will operate on a clean directory.

[endlessm/eos-shell#3053]