scripts: always ensure directory structure
The rollback/uninstall scripts call into delete_symbolic_links();
whenever we add a new directory to the list, if that directory wasn't
created when installing, perhaps because it was installed with an older
version of the scripts, it will currently fail trying to remove symlinks
from a non-existent location.

Just ensure the directory structure on any transaction instead.

[endlessm/eos-shell#3010]