Determine the app architecture at build time
We need to map to the Debian architecture from the host's architecture
details. This is a lot easier to do at build time where we have access
to the GNU triplet. Allow the user to set the architecture through
--with-eos-arch, but fallback to some heuristics for architectures that
we care about at Endless. This should actually do the right thing for
all but a few oddball architectures.

[endlessm/eos-shell#3161]