rest: Include arch in the server API queries
In order to support multiple architectures, the app manager needs to
query the server with the system architecture so it will receive
appropriate bundles. Unfortunately, arch was not in the original v1 API
and needs to be passed as a parameter in all cases except when the app
query is fully specified with the version.

[endlessm/eos-shell#3161]