Notify available apps changed after (un)installing an app
The list of available applications are formed by those applications that are in
the server but not installed in our system.

So if we install an application, it becomes unavailable. And the opposite, if
we uninstall an application now it becomes available.

Thus, when installing or uninstalling an application we need to notify that the
list of available applications has changed.

[endlessm/eos-shell#3158]