scripts: Use gpgv to verify bundle signature
gpg is intended as a user program and expects to use ~/.gnupg or other
pre-existing gpg store. gpgv is designed to do simple pgp signature
verification with a designated keyring and doesn't require any
pre-existing setup. Let's just use that. This is also how ostree does
signature verification for us.

[endlessm/eos-shell#3505]