scripts: use --ignore-euid for xdelta3-dir-patcher
Since the change of application-manager to run as non-root, the
xdelta3-dir-patcher stopped to work, since it, by default, expects to be
executed as root in order to keep the same file permissions.

This patch fix this issue by setting the parameter --ignore-euid , which seems
enough since the whole /endless filesystem is manager by the app-manager user.

[endlessm/eos-shell#3899]