updates: Don't filter returned apps
The server will already do all the complicated filtering based on the
parameters we sent it, so there's no need to do any filtering on the
client side. This breaks the case where the server sends back a blank
personality string, which is intended to mean that it's for any
personality.

[endlessm/eos-shell#4005]