eam-fs-sanity: fix permissions for application files when needed
We've seen cases where an installed bundle in /endless would not have
the right permissions set for its directory and its files (e.g. ri-li),
making it impossible for the user to use that application, as well as
for the app store to show it as an "installed application", since access
to the relevant files (e.g. binaries, .desktop files) would not be possible.

This patch performs a quick check when calling eam_fs_sanity_check()
that would cause EAM to try to restore permissions for the files of a
bundle to a more sane state when the app directory is not readable by
"others", which is needed for the app store to see them.

It's not a perfect "check and fix" algorithm, but it's quick enough not
to add a significant extra overhead on startup, and probably sane enough
as it considers previously present permission for the files/dirs it has
to fix, adding only the 'r-x' ones on top if, and only if, those
permissions are set for the owner too.

[endlessm/eos-shell#4250]