Added installed size values to DBus listings of available apps
This should provide the app store with the needed data to figure out if
we can install the app or not based on space requirements. Value used is
a signed long to have the option of using negative values as special
flags if needed (e.g. -1 = explicitly unknown expanded size).