fs-sanity: Create directory structure in appdir
When /endless is an overlayfs rather than a symlink, creation of missing
subdirectories in /var/endless triggers a bug where the newly created
directory can't be read through the overlay. The sanity check creates
the application subdirectories directly in the real app directory under
/var. This is unlike the installation scripts, which always operate
through the rooted path at /endless.

This is a bug in overlayfs, but refactor so that the subdirectory
structure is created in the appdir after the symlink is checked to
exist. This way all app installation paths operate through /endless
rather than directly manipulating the underlying directories.

[endlessm/eos-shell#5036]