service: do not double free remote transactions when cancelled
When we receive a call to CancelTransaction(), we can be in the
following two scenarios:
- Install()/Update() has been called, an object path has been returned
  to the app store, which hasn't called CompleteTransaction() yet.
- The app store has called CompleteTransaction(), and the app manager is
  performing the installation on disk.

The main difference is that in the latter case, a GCancellable regulates
the installation process on disk; when we call g_cancellable_cancel() in
that case, the app manager will also receive a callback (with error)
from the installation process, and we'll end up free-ing the remote
transaction twice.
Instead, avoid free-ing the remote transaction from the installation
process callback, as the code immediately following
g_cancellable_cancel() will do that already.

[endlessm/eos-shell#5167]