Fixed handling of leftover symlinks
If we didn't cleanup all symlinks from previous app install, the app
manager would choke and die but now we do a forced replacement. The
proper fix for this is to ensure that we don't leave any files behind
but at this point anyone that uninstalled apps with the old code would
have junk leftover so for now, this "forced removal" approach is
preferred.