fs: Don't create /endless symlink or /var/endless directory
The policy on how to handle the /endless app directory is up to the
local system and does not need to be enforced in code. The app manager
should only try to create the directories directly in /endless. This
would fail on systems where /endless in an overlayfs as well as systems
like the image builder where the migration to /var/endless will happen
after the apps are installed.

[endlessm/eos-shell#5257]