Remove allow_deltas flag from DBus interface
This does not make any sense here; the app store is the only component
in control of whether a specific update payload is a delta update or
not, and the app manager should not consult the configuration again.

We can just look at the nature of the file itself to determine whether
we should use the delta routines or not.