service: Allow specifying the target storage for installations
We want to allow callers of the app manager API to decide where an
application should be installed.

Instead of specifying the path, we allow specifying the storage type
(either 'primary' or 'secondary'), and then we resolve the right path
using the eos-app-manager configuration.

If the storage type is missing, we default to the primary one.

[endlessm/eos-shell#5018]