fs: Strip correct directory from PATH
When trying to look for the program to create a symlink to, we try to
take the /endless/bin directory out of the search path so that we don't
find an existing symlink there. However, the directory being removed
from the path was just the appdir relative "bin". Build the full correct
path to the directory.

[endlessm/eos-shell#5238]