Clean up all the configuration options
The configuration file is a mess of old options, and it's barely user
editable.

We need to regroup the options per scope; drop the unused options
related to the remote repository; clean up the options we still use in a
way that makes sense for what we use them today; and add new options
that we're going to use in the future.

[endlessm/eos-shell#5018]