fs-sanity: Symlink in /endless/bin through the binaries symlinks function
Instead of symlinking the binaries through the bundle dir, we now use
the special do_binaries_symlinks function as a special-case.

This was actually running before, but we forgot to make the filesystem
directory during ensure_symlinks and we forgot to skip the bindir when
recursively symlinking files during installation, so it was as if the
logic wasn't there.

[endlessm/eos-shell#5414]