fs-sanity: Don't ever skip symlinking apps
Previously, since we were running this on every boot, we implemented an optimization to skip apps that seem to already exist.

Recently, we realized this is racy: it's possible for the symlink of the app dir to get created, but not have a "full deployment" of all the symlinks. Now that we're using systemd's upgrade system to run this script, just remove this optimization.