remote-transaction: fix a segfault when cancelling an operation
When CancelTransaction() is called on a transaction exported on the bus,
the app manager will cancel the GCancellable associated with it, and
free the transaction, removing it from the bus.
But if the transaction is in an asynchronous operation, such as
those triggered by eam_transaction_run_async(), its callback will
eventually still be called, with eam_transaction_run_finish() returning
a GError.
In order to call eam_transaction_run_finish(), code currently accesses
the EamRemoteTransaction structure, which was though already released
with CancelTransaction() as per above.

This commit changes the code so that when the operation is cancelled,
the callback of eam_transaction_run_async() does not unpack
EamRemoteTransaction to call eam_transaction_run_finish(), and returns
immediately instead.

To make this possible, eam_remote_transaction_cancel() will now complete
(with error) the GDBusMethodInvocation of CompleteTransaction(), if
there is one.

[endlessm/eos-shell#5923]