Allow moving between storages on update
When updating an application we end up requiring more space because of
the backup we use for rolling back on error conditions.

This means that the update transation needs to be able to use a source
prefix to find the application where it is installed; and it needs a
target prefix, to define the location where the application should be
installed.

In order to allow this behaviour, we need to change the transaction's
DBus API, by adding new keys to the options.

[endlessm/eos-shell#6136]