eam-utils: Silence stdout
On successful runs of eamctl install, we will otherwise see the
confusing error message: "No schema files, doing nothing". This is 100%
normal and intended and comes from glib-compile-schemas, but is
understandably confusing to people not familiar with the process.

Let's just kill stdout -- any useful errors should be in stderr
instead.

https://phabricator.endlessm.com/T11879