Revert "migrate-bundles: Ensure app locations are mounted"
This was breaking single-disk systems.

Per Dan Nicholson, the RequiresMountsFor is telling systemd
that the mount unit is now required. It tries to start the
mount unit and fails because there's no extra filesystem.
Then it doesn't run the actual migrate-bundles service because
a requirement failed. The real issue was fixed in eos-boot-helper
with the SD card mount ordering. Now the migrate-bundles service
will work fine because the SD card will be reliably mounted
before local-gs.target.

This reverts commit 9a52cc3fa7d6b5c730c6e9427dd8d6ea80960687.